ParserWrapper in a separate file
